# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AddModel": {
        "properties": {
          "api_key_env": {
            "default": "",
            "title": "Api Key Env",
            "type": "string"
          },
          "base_url": {
            "default": "",
            "title": "Base Url",
            "type": "string"
          },
          "context": {
            "default": 128000,
            "exclusiveMinimum": 0.0,
            "title": "Context",
            "type": "integer"
          },
          "input_per_m": {
            "default": 0.0,
            "title": "Input Per M",
            "type": "number"
          },
          "json_mode": {
            "default": true,
            "title": "Json Mode",
            "type": "boolean"
          },
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_per_m": {
            "default": 0.0,
            "title": "Output Per M",
            "type": "number"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model"
        ],
        "title": "AddModel",
        "type": "object"
      },
...
      "BulkRoute": {
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "model"
        ],
        "title": "BulkRoute",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RouteUpdate": {
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          },
          "task": {
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task",
          "model"
        ],
        "title": "RouteUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Studio Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/models/catalogue": {
      "post": {
        "description": "Register any model \u2014 a new Ollama pull, an OpenAI-compatible gateway, whatever.\n\nThe catalogue is a starting point, not a whitelist.",
        "operationId": "add_model_v1_models_catalogue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Model V1 Models Catalogue Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Model",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/ollama": {
      "get": {
        "description": "What Ollama actually has installed, so the UI offers real models.",
        "operationId": "ollama_v1_models_ollama_get",
        "parameters": [
          {
            "in": "query",
            "name": "base_url",
            "required": false,
            "schema": {
              "default": "http://localhost:11434",
              "title": "Base Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ollama V1 Models Ollama Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ollama",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/ollama/register": {
      "post": {
        "description": "Add every installed Ollama model to the catalogue in one go.\n\n`json_mode` is set true because Ollama constrains decoding with `format: json`,\nwhich is what makes small local models usable for the structured stages. It is\nnot a promise the output will be *good*.",
        "operationId": "register_ollama_v1_models_ollama_register_post",
        "parameters": [
          {
            "in": "query",
            "name": "base_url",
            "required": false,
            "schema": {
              "default": "http://localhost:11434",
              "title": "Base Url",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Register Ollama V1 Models Ollama Register Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Ollama",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/route": {
      "put": {
        "operationId": "set_route_v1_models_route_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set Route V1 Models Route Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Route",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/route/all": {
      "put": {
        "description": "Route every task to one model \u2014 the 'run it all locally' button.",
        "operationId": "set_all_v1_models_route_all_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRoute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set All V1 Models Route All Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set All",
        "tags": [
          "models"
        ]
      }
    },
...
    "/v1/models/test": {
      "post": {
        "description": "Round-trip one model so a broken route is found here, not mid-render.",
        "operationId": "test_model_v1_models_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRoute"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Test Model V1 Models Test Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Model",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}